feat: add --lean-version-tag parameter for offline sorrys
The version tag parameter allows users to specify the lean version that
REPL uses that the version of the Lean project match the
version of REPL.

Modify `src/repl_api.py` to checkout the REPL repo at this version.

The default is still the most recent version of REPL available.