feat: add build_sorry_db script
build_sorry_db script calls the new `build_database` method which
loops through a list of repos in the format of `test_repos.json`
and creates a json database from all the sorries in the repo.

The database is a flat structure, i.e., each sorry record duplicates
the repo metadata.

`build_database` also generates a uuid for each sorry.

Note, the name of the build_sorry_db scripts is intentionally different
from the `build_database` method, but we may should align them.