chore: update the Intro Jupyter notebook
This makes a few minor tweaks to the doc notebook
Those include:

    * using `pathlib` in a few more places
    * trimming `https://github.com` when printing out repositories,
        to perhaps make the output a bit less noisy
    * noting that `elan` is installed by one of the cells

And then we re-run the notebook to regenerate its output.

(Note there's a single cell I didn't run here, the one which installs
elan, as I've already got it installed, but it seems reasonable to leave
that output out of the notebook).